#376 Fix Windows tests
